requirements.txt
openai>=1.0.0
requests>=2.31.0
feedparser>=6.0.10
google-auth>=2.23.0
google-auth-oauthlib>=1.1.0
google-auth-httplib2>=0.1.1
google-api-python-client>=2.100.0
python-dotenv>=1.0.0
pydantic>=2.0.0
httpx>=0.24.0
python-dateutil>=2.8.2
